Move tasks to new package
