+++ src/main/java/io/github/cccm5/CargoMain.java
@@ -7,6 +7,9 @@ import com.degitise.minevid.dtlTraders.guis.gui.TradeGUIPage;
 import com.degitise.minevid.dtlTraders.guis.items.AGUIItem;
 import com.degitise.minevid.dtlTraders.guis.items.TradableGUIItem;
 import com.degitise.minevid.dtlTraders.utils.citizens.TraderTrait;
+import io.github.cccm5.async.LoadTask;
+import io.github.cccm5.async.ProcessingTask;
+import io.github.cccm5.async.UnloadTask;
 import io.github.cccm5.util.CraftInventoryUtil;
 import io.github.cccm5.util.NPCUtil;
 import net.citizensnpcs.api.CitizensAPI;
