+++ src/main/java/io/github/cccm5/async/LoadTask.java
@@ -1,6 +1,7 @@
-package io.github.cccm5;
+package io.github.cccm5.async;
 
 import com.degitise.minevid.dtlTraders.guis.items.TradableGUIItem;
+import io.github.cccm5.CargoMain;
 import io.github.cccm5.util.CraftInventoryUtil;
 import net.countercraft.movecraft.craft.PlayerCraft;
